Dragonflight updates

+++ Bags.lua
@@ -11,9 +11,9 @@ function Module:OnInitialize()
 end
 
 function enableGlow(self)
-    local quality = select(4, GetContainerItemInfo(self:GetParent():GetID(), self:GetID()))
+    local item = C_Container.GetContainerItemInfo(self:GetParent():GetID(), self:GetID())
 
-    if quality and quality > Enum.ItemQuality.Common then
+    if item and item.quality > Enum.ItemQuality.Common then
         self.NewItemTexture:SetAtlas(NEW_ITEM_ATLAS_BY_QUALITY[quality])
         self.NewItemTexture:SetAlpha(.8)
         self.NewItemTexture:Show()
@@ -66,22 +66,22 @@ function Module:DestroyItems()
 
     -- LuaFormatter on
     for bagID = BACKPACK_CONTAINER, NUM_BAG_SLOTS do
-        for slot = 1, GetContainerNumSlots(bagID) do
-            local itemLink = GetContainerItemLink(bagID, slot) or ''
+        for slot = 1, C_Container.GetContainerNumSlots(bagID) do
+            local itemLink = C_Container.GetContainerItemLink(bagID, slot) or ''
 
             local itemName = string.match(itemLink, '|h%[([^%]]+)%]|h')
-            local quality = select(4, GetContainerItemInfo(bagID, slot))
+            local item = C_Container.GetContainerItemInfo(bagID, slot)
 
             -- check if item is an heirloom
-            if quality == 7 and not protectedItems[itemName] then
-                PickupContainerItem(bagID, slot)
+            if item and item.quality == 7 and not protectedItems[itemName] then
+                C_Container.PickupContainerItem(bagID, slot)
                 DeleteCursorItem()
             end
 
             -- check if item in on the destroy list
             for _, destroyItemName in ipairs(destroyItems) do
                 if itemName == destroyItemName then
-                    PickupContainerItem(bagID, slot)
+                    C_Container.PickupContainerItem(bagID, slot)
                     DeleteCursorItem()
                 end
             end
